Route MoE execution through kernels API
